Create README.md
The dataset is a sample of a brand value study for a brand that sells a product. The first set of variables are demographics and a set of brand value or brand imagery statements (asked on a likert scale 1 to 5). The respondents were then asked how likely they were to purchase the company's product on a 1-10 scale. The task is to know what drives the likelihood of purchase